refactor: 바인딩 어댑터 패키지 정리
